Add console log to display cookies in Home function
A console.log statement has been added to the Home function to display cookies. This is useful for testing and debugging purposes to understand the cookies being used or available at that point in the application. The output of this console.log can be viewed in the browser's developer console.